--- buklod-tao-branch/editloc.html
@@ -1,149 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>Edit Location</title>
-    <link rel="stylesheet" href="./css/form.css">
-
-    <script type="module">
-        import { editLocation, getDocIdByPartnerName, getDocByID } from "./js/firestore.js";
-        function populateFormFields(partnerData) {
-            document.getElementById("partner").value = partnerData.name;
-            document.getElementById("partner_contact-person").value = partnerData["`partner-contact`"];
-            document.getElementById("organization").value = partnerData.org;
-            document.getElementById("nature_of_act").value = partnerData.activity;
-            document.getElementById("dates_of_partnership").value = partnerData.dates;
-            document.getElementById("ateneo_office").value = partnerData["`admu-office`"];
-            document.getElementById("ateneo_contact-person").value = partnerData["`admu-contact`"];
-            document.getElementById("ateneo_contact-email").value = partnerData["`admu-email`"];
-        }
-        function fetchPartnerData() {
-            const params = new URLSearchParams(window.location.search);
-            const partnerName = params.get("partnerName");
-            if (partnerName) {
-                getDocIdByPartnerName(partnerName)
-                    .then((docId) => {
-                        if (docId) {
-                            getDocByID(docId).then((doc) => {
-                                populateFormFields(doc);
-                            });
-                        }
-                        else {
-                            alert("No matching document found for the partner name: " + partnerName);
-                        }
-                    });
-            }
-            else {
-                alert("No partnerName parameter found in the URL.");
-            }
-        }
-        fetchPartnerData();
-
-        document.getElementById("submit_form").addEventListener("click", handleFormSubmit);
-
-        function handleFormSubmit() {
-            const partnerName = document.getElementById("partner").value;
-            getDocIdByPartnerName(partnerName)
-                .then((docId) => {
-                    if (docId) {
-                        // If a matching docId is found, proceed with updating the document
-                        const name = partnerName;
-                        const activity = document.getElementById("nature_of_act").value;
-                        const admuContact = document.getElementById("ateneo_contact-person").value;
-                        const admuEmail = document.getElementById("ateneo_contact-email").value;
-                        const admuOffice = document.getElementById("ateneo_office").value;
-                        const org = document.getElementById("organization").value;
-                        const partnerContact = document.getElementById("partner_contact-person").value;
-                        const dates = document.getElementById("dates_of_partnership").value;
-
-                        editLocation(docId, name, activity, admuContact, admuEmail, admuOffice, org, partnerContact, dates);
-                    }
-                    else {
-                        // Handle case when no matching document is found
-                        alert("No matching document found for the partner name: " + partnerName);
-                    }
-                });
-
-        }
-    </script>
-</head>
-
-<body>
-    <br>
-    <div class="container">
-        <!-- Div for margin -->
-        <h1>Edit Location</h1>
-        <br>
-
-        <!-- Form -->
-        <div class="input_container">
-            <label class="form" for="partner">Partner</label><br>
-            <input type="text" id="partner" name="partner" class="attribs" placeholder="">
-        </div>
-
-        <br>
-
-        <div class="input_container">
-            <label class="form" for="partner_contact-person">Contact Person of Partner</label><br>
-            <input type="text" id="partner_contact-person" name="partner_contact-person" class="attribs" placeholder="">
-        </div>
-
-        <br>
-
-        <div class="input_container">
-            <label class="form" for="organization">Organization / Unit</label><br>
-            <input type="text" id="organization" name="organization" class="attribs" placeholder="">
-        </div>
-
-        <br>
-
-        <div class="input_container">
-            <label class="form" for="nature_of_act">Nature of Activity</label><br>
-            <input type="text" id="nature_of_act" name="nature_of_act" class="attribs" placeholder="">
-        </div>
-
-        <br>
-
-        <!-- Pressing a button adds another date -->
-
-        <div class="input_container">
-            <label class="form" for="dates_of_partnership">Dates of Partnership</label><br>
-            <input type="date" id="dates_of_partnership" name="dates_of_partnership" class="attribs" placeholder="">
-        </div>
-
-        <br>
-
-        <div class="input_container">
-            <label class="form" for="ateneo_office">Ateneo Office Oversight</label><br>
-            <input type="text" id="ateneo_office" name="ateneo_office" class="attribs" placeholder="">
-        </div>
-
-        <br>
-
-        <div class="input_container">
-            <label class="form" for="ateneo_contact-person">Ateneo Contact Person</label><br>
-            <input type="text" id="ateneo_contact-person" name="ateneo_contact-person" class="attribs" placeholder="">
-        </div>
-
-        <br>
-
-        <div class="input_container">
-            <label class="form" for="ateneo_contact-email">Email</label><br>
-            <input type="email" id="ateneo_contact-email" name="ateneo_contact-email" class="attribs" placeholder="">
-        </div>
-
-        <br><br>
-
-        <!-- Buttons -->
-        <span class="button_div">
-            <!-- Clears all contents of the form -->
-            <button class="clear form">Clear Contents</button>
-            <!-- Submits contents of the form -->
-            <button class="confirm form" type="submit" id="submit_form" name="submit_form">Save Edit</button>
-        </span>
-
-        <br><br><br>
-    </div>
-</body>
-
-</html>
