--- editloc.html
@@ -1,279 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<title>Edit Location</title>
-
-		<!-- Import Tailwind CSS output file -->
-		<link href="./src/output.css" rel="stylesheet" />
-
-		<!-- Import Fonts -->
-		<link rel="preconnect" href="https://fonts.googleapis.com" />
-		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-		<link
-			href="https://fonts.googleapis.com/css2?family=EB+Garamond:ital,wght@0,400..800;1,400..800&display=swap"
-			rel="stylesheet"
-		/>
-		<link rel="preconnect" href="https://fonts.googleapis.com" />
-		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-		<link
-			href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap"
-			rel="stylesheet"
-		/>
-		<link rel="preconnect" href="https://fonts.googleapis.com" />
-		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-		<link
-			href="https://fonts.googleapis.com/css2?family=Hind:wght@300;400;500;600;700&display=swap"
-			rel="stylesheet"
-		/>
-
-		<script type="module">
-			import {
-				editLocation,
-				getDocIdByPartnerName,
-				getDocByID,
-			} from './js/firestore.js';
-			function populateFormFields(partnerData) {
-				document.getElementById('partner').value = partnerData.name;
-				document.getElementById('partner_contact-person').value =
-					partnerData['`partner-contact`'];
-				document.getElementById('organization').value =
-					partnerData.org;
-				document.getElementById('nature_of_act').value =
-					partnerData.activity;
-				document.getElementById('dates_of_partnership').value =
-					partnerData.dates;
-				document.getElementById('ateneo_office').value =
-					partnerData['`admu-office`'];
-				document.getElementById('ateneo_contact-person').value =
-					partnerData['`admu-contact`'];
-				document.getElementById('ateneo_contact-email').value =
-					partnerData['`admu-email`'];
-			}
-			function fetchPartnerData() {
-				const params = new URLSearchParams(window.location.search);
-				const partnerName = params.get('partnerName');
-				if (partnerName) {
-					getDocIdByPartnerName(partnerName).then((docId) => {
-						if (docId) {
-							getDocByID(docId).then((doc) => {
-								populateFormFields(doc);
-							});
-						} else {
-							alert(
-								'No matching document found for the partner name: ' +
-									partnerName
-							);
-						}
-					});
-				} else {
-					alert('No partnerName parameter found in the URL.');
-				}
-			}
-			fetchPartnerData();
-
-			document
-				.getElementById('submit_form')
-				.addEventListener('click', handleFormSubmit);
-
-			function handleFormSubmit() {
-				const partnerName =
-					document.getElementById('partner').value;
-				getDocIdByPartnerName(partnerName).then((docId) => {
-					if (docId) {
-						// If a matching docId is found, proceed with updating the document
-						const name = partnerName;
-						const activity =
-							document.getElementById(
-								'nature_of_act'
-							).value;
-						const admuContact = document.getElementById(
-							'ateneo_contact-person'
-						).value;
-						const admuEmail = document.getElementById(
-							'ateneo_contact-email'
-						).value;
-						const admuOffice =
-							document.getElementById(
-								'ateneo_office'
-							).value;
-						const org =
-							document.getElementById(
-								'organization'
-							).value;
-						const partnerContact = document.getElementById(
-							'partner_contact-person'
-						).value;
-						const dates = document.getElementById(
-							'dates_of_partnership'
-						).value;
-
-						editLocation(
-							docId,
-							name,
-							activity,
-							admuContact,
-							admuEmail,
-							admuOffice,
-							org,
-							partnerContact,
-							dates
-						);
-					} else {
-						// Handle case when no matching document is found
-						alert(
-							'No matching document found for the partner name: ' +
-								partnerName
-						);
-					}
-				});
-			}
-		</script>
-	</head>
-
-	<body>
-		<form action="" class="p-8 bg-customLightGray w-1/2">
-			<!-- Div for margin -->
-			<h1 class="font-montserrat text-3xl font-bold text-darkbg mb-8">
-				Edit Activity
-			</h1>
-
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="partner"
-					>Partner</label
-				><br />
-				<input
-					type="text"
-					id="partner"
-					name="partner"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="partner_contact-person"
-					>Contact Person of Partner</label
-				>
-				<input
-					type="text"
-					id="partner_contact-person"
-					name="partner_contact-person"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="organization"
-					>Organization / Unit</label
-				><br />
-				<input
-					type="text"
-					id="organization"
-					name="organization"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="nature_of_act"
-					>Nature of Activity</label
-				><br />
-				<input
-					type="text"
-					id="nature_of_act"
-					name="nature_of_act"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<!-- TODO: Add a button for adding another date -->
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="dates_of_partnership"
-					>Dates of Partnership</label
-				><br />
-				<input
-					type="date"
-					id="dates_of_partnership"
-					name="dates_of_partnership"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="ateneo_office"
-					>Ateneo Office Oversight</label
-				><br />
-				<input
-					type="text"
-					id="ateneo_office"
-					name="ateneo_office"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="ateneo_contact-person"
-					>Ateneo Contact Person</label
-				><br />
-				<input
-					type="text"
-					id="ateneo_contact-person"
-					name="ateneo_contact-person"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<div class="flex flex-row justify-between mb-6 items-center">
-				<label
-					class="form font-montserrat text-xl text-customGray"
-					for="ateneo_contact-email"
-					>Email</label
-				><br />
-				<input
-					type="email"
-					id="ateneo_contact-email"
-					name="ateneo_contact-email"
-					class="py-4 px-8 w-96 rounded border font-montserrat"
-					placeholder=""
-				/>
-			</div>
-
-			<!-- Buttons -->
-			<div class="">
-				<button
-					class="w-full py-4 text-center font-montserrat text-2xl bg-lightbg text-white font-bold rounded-lg"
-					id="submit_form"
-					name="submit_form"
-					type="submit"
-				>
-					Save
-				</button>
-				<button
-					class="w-full py-4 text-center font-montserrat text-2xl text-red-700 font-bold border border-red-700 rounded-lg mt-4"
-				>
-					Delete
-				</button>
-			</div>
-		</form>
-	</body>
-</html>
